LayoutTests/media/video-load-networkState-expected.txt

EXPECTED (video.networkState == '0') OK
EVENT(loadstart)
EXPECTED (video.networkState == '1') OK
EVENT(loadedmetadata)
EXPECTED (video.networkState == '2') OK
EVENT(loadedfirstframe)
EXPECTED (video.networkState == '3') OK
EVENT(load)
EXPECTED (video.networkState == '4') OK
END OF TEST
